chore: Add Dockerfile and update dependencies
Introduced a Dockerfile for containerized builds and production deployment. Updated dependencies in package.json and package-lock.json, including addition of 'pg' for PostgreSQL support and upgrades to NestJS and related packages.

+++ src/balance-detail/balance-detail.entity.ts
@@ -41,6 +41,6 @@ export class BalanceDetailEntity {
   @Column({ name: 'cash', type: 'int' })
   cash: number;
 
-  @Column({ name: 'change', type: 'double' })
+  @Column({ name: 'change', type: 'float' })
   change: number;
 }
