+++ src/balance/balance.entity.ts
@@ -24,13 +24,13 @@ export class BalanceEntity {
   @Column({ name: 'time', type: 'time' })
   time: Date;
 
-  @Column({ name: 'initial_balance', type: 'double' })
+  @Column({ name: 'initial_balance', type: 'float' })
   initialBalance: number;
 
-  @Column({ name: 'final_system_balance', type: 'double' })
+  @Column({ name: 'final_system_balance', type: 'float' })
   finalSystemBalance: number;
 
-  @Column({ name: 'final_user_balance', type: 'double' })
+  @Column({ name: 'final_user_balance', type: 'float' })
   finalUserBalance: number;
 
   @OneToMany(
